+++ palm/src/jasmine/mod.rs
@@ -10,7 +10,7 @@ use self::protocols::{S3SyncClient, TS3SyncClient};
 const S3_SERVICE_NAME: &str = "github.com/saturn-xiv/palm/jasmine/services/v1/S3";
 
 pub trait S3 {
-    fn create_bucket(&self, name: &str, public: bool, expiration_days: i8) -> ThriftResult<()>;
+    fn create_bucket(&self, name: &str, public: bool, expiration_days: i32) -> ThriftResult<()>;
     fn upload_file(&self, bucket: &str, object: &str, ttl: Duration) -> ThriftResult<String>;
     fn get_presigned_url(
         &self,
@@ -22,7 +22,7 @@ pub trait S3 {
     fn get_permanent_url(&self, bucket: &str, object: &str) -> ThriftResult<String>;
 }
 impl S3 for Thrift {
-    fn create_bucket(&self, name: &str, public: bool, expiration_days: i8) -> ThriftResult<()> {
+    fn create_bucket(&self, name: &str, public: bool, expiration_days: i32) -> ThriftResult<()> {
         let (i_prot, o_prot) = self.open(S3_SERVICE_NAME)?;
         let mut client: S3SyncClient<_, _> = S3SyncClient::new(i_prot, o_prot);
         client.create_bucket(name.to_string(), public, expiration_days)?;
